requirements.txt
addict==2.4.0
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164153/work
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1656782824682/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
commonmark==0.9.1
conda==4.14.0
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1663583595612/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1657174000558/work
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.37.3
gitdb==4.0.9
GitPython==3.1.27
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1661285697648/work/libmambapy
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1661285697648/work/mamba
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mmcv-full==1.6.1
mmdet==2.25.1
model-index==0.1.11
numpy==1.21.6
opencv-python==4.6.0.66
openmim==0.3.2
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyarrow==9.0.0
pycocotools==2.0.5
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1649384814992/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyngrok==5.1.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python==3.7.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
rich==12.5.1
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1653464404698/work
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
tabulate==0.8.10
terminaltables==3.1.10
toml==0.10.2
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
validators==0.20.0
watchdog==2.1.9
yapf==0.32.0
zipp==3.8.1
